requirements.txt
FastAPI
azure-cognitiveservices-speech
pydantic
uvicorn
python-dotenv
requests